requirements.txt
requests
redis
boto3
psycopg2-binary
